+++ Source/threecs_project/Private/Character/Logic/MyCharacter.cpp
@@ -1,7 +1,7 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 
-#include "MyCharacter.h"
+#include "Character/Logic/MyCharacter.h"
 #include "GameFramework/SpringArmComponent.h"
 
 // Sets default values
